requirements.txt
# ============================================================
# English Buddy Backend - Dependencies
# ============================================================
flask==3.0.*
flask-cors==5.0.*
requests==2.31.*
python-dotenv==1.0.*
gunicorn==22.0.*
